enh: Custom fmt chunk #49
Improved parsing of WAVE file header. It should only get the information we need, regardless of number of chunks, and ignore other chunks.
Previously parsing was limited to fixed, 44 byte header. Improved parsing so it checks for the correct chunk and fetches only data that is needed.
No more fixed "fmt " chunk size.